+++ R/class-spec.R
@@ -80,20 +80,74 @@ class_friendly <- function(x) {
   )
 }
 
-class_constructor <- function(.x, ...) {
-  switch(class_type(.x),
-    NULL = function() NULL,
-    any = function() NULL,
-    S4 = function(...) methods::new(.x, ...),
-    S7 = .x,
-    S7_base = .x$constructor,
-    S7_union = class_constructor(.x$classes[[1]]),
-    S7_S3 = .x$constructor,
-    stop(sprintf("Can't construct %s", class_friendly(.x)), call. = FALSE)
-  )
-}
 class_construct <- function(.x, ...) {
-  class_constructor(.x)(...)
+  eval(class_construct_expr(.x, ...))
+}
+
+
+class_construct_expr <- function(.x, ...) {
+  f <- class_constructor(.x)
+  # If the constructor is a closure wrapping a simple expression, try
+  # to extract the expression
+  # (mostly for nicer printing and introspection.)
+
+  ## early return if not safe to unwrap
+  # can't unwrap if we're passing on ...
+  if(...length()) {
+    return(as.call(list(f, ...)))
+  }
+
+  # can't unwrap if the closure is potentially important
+  # (this can probably be relaxed to allow additional environments)
+  fe <- environment(f)
+  if(!identical(fe, baseenv())) {
+    return(as.call(list(f, ...)))
+  }
+
+  # special case for `class_missing`
+  if (identical(body(f) -> fb, quote(expr =))) {
+    return(quote(expr =))
+  }
+
+  # `new_object()` must be called from the class constructor, can't
+  # be safely unwrapped
+  if("new_object" %in% all.names(fb)) {
+    return(as.call(list(f, ...)))
+  }
+
+  # maybe unwrap body if it is a single expression wrapped in `{`
+  if (length(fb) == 2L && identical(fb[[1L]], quote(`{`)))
+    fb <- fb[[2L]]
+
+  # If all the all the work happens in the promise to the `.data` arg,
+  # return the `.data` expression.
+  ff <- formals(f)
+  if ((identical(fb, quote(.data))) &&
+      identical(names(ff), ".data")) {
+    return(ff$.data)
+  }
+
+  # if all the work happens in the function body, return the body.
+  if (is.null(ff)) {
+    return(fb)
+  }
+
+  #else, return a call to the constructor
+  as.call(list(f, ...))
+}
+
+class_constructor <- function(.x) {
+  switch(class_type(.x),
+         NULL = function() NULL,
+         missing = new_function(, quote(expr =), baseenv()),
+         any = function() NULL,
+         S4 = function(...) methods::new(.x, ...),
+         S7 = .x,
+         S7_base = .x$constructor,
+         S7_union = class_constructor(.x$classes[[1]]),
+         S7_S3 = .x$constructor,
+         stop(sprintf("Can't construct %s", class_friendly(.x)), call. = FALSE)
+  )
 }
 
 class_validate <- function(class, object) {